Add parallel π computation using OpenMP in P3.cpp
A new function `compute_pi_in_parallel()` has been added to P3.cpp to compute the value of π in parallel using OpenMP. This function initializes several variables and defines a constant `THREAD_SIZE` set to 16. It calculates the `step` variable, records the start time, and creates an OpenMP parallel region with `THREAD_SIZE` threads. Each thread calculates a local sum for its portion of the iterations, and the value of π is computed by multiplying the `step` by the total `sum`. The function records the run time and prints the result. The `main()` function has been updated to call this new function after the existing `compute_pi()` function. Additionally, several commented-out lines of code have been added to the `main()` function.

Add parallel pi computation with OpenMP in P3.cpp

A new function `compute_pi_in_parallel()` has been added to the
`P3.cpp` file. This function calculates the value of pi in parallel
using OpenMP, initializes variables, sets the number of threads,
and uses a parallel region to divide the computation among multiple
threads. The result is then printed along with the computation time.

The `main()` function has been updated to call the new
`compute_pi_in_parallel()` function after calling the existing
`compute_pi()` function.

Several lines of commented-out code have been added to the `main()`
function, which appear to be related to some iterative calculations
and phases, but these lines are currently not active in the code.